+++ demo/src/client/faro/initialize.ts
@@ -1,5 +1,6 @@
 import { createRoutesFromChildren, matchRoutes, Routes, useLocation, useNavigationType } from 'react-router-dom';
 
+import { XHRInstrumentation } from '@grafana/faro-instrumentation-xhr';
 import {
   initializeFaro as coreInit,
   getWebInstrumentations,
@@ -32,6 +33,7 @@ export function initializeFaro(): Faro {
           },
         },
       }),
+      new XHRInstrumentation(),
     ],
     batching: {
       // Batching is enabled by default and there is normally no reason to disable it.