+++ experimental/instrumentation-fetch/src/instrumentation.test.ts
@@ -80,6 +80,74 @@ describe('FetchInstrumentation', () => {
       request: {
         method: 'GET',
         url: 'https://example.com/',
+        headers: {},
+        destination: '',
+        referrerPolicy: '',
+        mode: 'cors',
+        credentials: 'same-origin',
+        cache: 'default',
+        redirect: 'follow',
+        integrity: '',
+        bodyUsed: false,
+      },
+    };
+
+    expect(actualResult.init).toStrictEqual(expectedResult.init);
+    expect(JSON.stringify(actualResult.request)).toEqual(JSON.stringify(expectedResult.request));
+  });
+
+  it('Adds RUM headers to same origin request', () => {
+    const instrumentation = new FetchInstrumentation({
+      testing: true,
+    });
+
+    const sessionInstrumentation = new SessionInstrumentation();
+
+    const faroInstance = initializeFaro(
+      makeCoreConfig(mockConfig({ instrumentations: [instrumentation, sessionInstrumentation] }))!
+    );
+    const sessionId = faroInstance.api.getSession()!.id as string;
+    expect(sessionId).not.toBe('');
+
+    const parseActualResult = (res: { init?: RequestInit | undefined; request: Request }) => {
+      return {
+        init: {
+          method: res.init?.method,
+          cache: res.init?.cache,
+        },
+        request: {
+          method: res.request.method,
+          url: res.request.url,
+          headers: res.request.headers,
+          destination: res.request.destination,
+          referrerPolicy: res.request.referrerPolicy,
+          mode: res.request.mode,
+          credentials: res.request.credentials,
+          cache: res.request.cache,
+          redirect: res.request.redirect,
+          integrity: res.request.integrity,
+          bodyUsed: res.request.bodyUsed,
+        },
+      };
+    };
+
+    const requestUrl = window.location.origin + '/test';
+
+    const actualResult = parseActualResult(
+      instrumentation.buildRequestAndInit(new Request(requestUrl), {
+        method: 'GET',
+        cache: 'no-cache',
+      })
+    );
+
+    const expectedResult = {
+      init: {
+        method: 'GET',
+        cache: 'no-cache',
+      },
+      request: {
+        method: 'GET',
+        url: requestUrl,
         headers: new Headers({
           'x-faro-session': makeFaroRumHeaderValue(sessionId),
         }),
@@ -94,8 +162,81 @@ describe('FetchInstrumentation', () => {
       },
     };
 
-    expect(actualResult.init?.toString()).toEqual(expectedResult.init?.toString());
-    expect(actualResult.request?.toString()).toEqual(expectedResult.request?.toString());
+    expect(actualResult.init).toStrictEqual(expectedResult.init);
+    expect({ ...actualResult.request, headers: actualResult.request.headers.get('x-faro-session') }).toStrictEqual({
+      ...expectedResult.request,
+      headers: expectedResult.request.headers.get('x-faro-session'),
+    });
+    expect(actualResult.request.headers.get('x-faro-session')).toBe(makeFaroRumHeaderValue(sessionId));
+  });
+
+  it('Adds RUM headers to requests to a different origin, which is explicitly allowed via the config', () => {
+    const instrumentation = new FetchInstrumentation({
+      testing: true,
+      propagateRumHeaderCorsUrls: ['https://example2.com'],
+    });
+
+    const sessionInstrumentation = new SessionInstrumentation();
+
+    const faroInstance = initializeFaro(
+      makeCoreConfig(mockConfig({ instrumentations: [instrumentation, sessionInstrumentation] }))!
+    );
+    const sessionId = faroInstance.api.getSession()!.id as string;
+    expect(sessionId).not.toBe('');
+
+    const parseActualResult = (res: { init?: RequestInit | undefined; request: Request }) => {
+      return {
+        init: {
+          method: res.init?.method,
+          cache: res.init?.cache,
+        },
+        request: {
+          method: res.request.method,
+          url: res.request.url,
+          headers: res.request.headers,
+          destination: res.request.destination,
+          referrerPolicy: res.request.referrerPolicy,
+          mode: res.request.mode,
+          credentials: res.request.credentials,
+          cache: res.request.cache,
+          redirect: res.request.redirect,
+          integrity: res.request.integrity,
+          bodyUsed: res.request.bodyUsed,
+        },
+      };
+    };
+    const actualResult = parseActualResult(
+      instrumentation.buildRequestAndInit(new Request('https://example2.com'), { method: 'GET', cache: 'no-cache' })
+    );
+
+    const expectedResult = {
+      init: {
+        method: 'GET',
+        cache: 'no-cache',
+      },
+      request: {
+        method: 'GET',
+        url: 'https://example2.com/',
+        headers: new Headers({
+          'x-faro-session': makeFaroRumHeaderValue(sessionId),
+        }),
+        destination: '',
+        referrerPolicy: '',
+        mode: 'cors',
+        credentials: 'same-origin',
+        cache: 'default',
+        redirect: 'follow',
+        integrity: '',
+        bodyUsed: false,
+      },
+    };
+
+    expect(actualResult.init).toStrictEqual(expectedResult.init);
+    expect({ ...actualResult.request, headers: actualResult.request.headers.get('x-faro-session') }).toStrictEqual({
+      ...expectedResult.request,
+      headers: expectedResult.request.headers.get('x-faro-session'),
+    });
+    expect(actualResult.request.headers.get('x-faro-session')).toBe(makeFaroRumHeaderValue(sessionId));
   });
 
   it('initialize FetchInstrumentation and calls fetch', () => {