+++ experimental/instrumentation-fetch/src/types.ts
@@ -1,3 +1,5 @@
+import type { Patterns } from '@grafana/faro-core';
+
 /**
  * Interface used to provide information to finish span on fetch error
  */
@@ -11,4 +13,10 @@ export interface FetchInstrumentationOptions {
   ignoredUrls?: Array<string | RegExp>;
   // For testing purposes - if true, fetch will be writable - necessary for jest tests
   testing?: boolean;
+
+  /**
+   * RUM headers are only added to URLs which have the same origin as the document.
+   * Ad other URLs which should have RUM headers added to this list.
+   */
+  propagateRumHeaderCorsUrls?: Patterns;
 }