+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Next
 
 - fix (`@grafana/faro-web-sdk`): guard against missing `isSampled` (#425)
+- fix (`@grafana/faro-instrumentation-fetch`): only add custom headers to requests sent to the same
+  origin as the document (#427)
+- fix (`@grafana/faro-instrumentation-xhr`): only add custom headers to requests sent to the same
+  origin as teh document (#427)
 
 ## 1.3.4